Consolidate option tests
