+++ java-tests/validating/src/test/kotlin/io/spine/test/options/setonce/SetOnceITest.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.test.setonce
+package io.spine.test.options.setonce
 
 import io.spine.test.tools.validate.name
 import io.spine.test.tools.validate.setOnceExplicitFalse
